Add manual assignment of apex
The coordinate apex of the heart can now be set with the -x, -y and -z
flags. The vertex in the mesh closest to the prescribed coordinates is
then set to be the apex.

Setting boundary conditions on a single vertex is not straight forward
in MFEM. Therefore, we set the apex to be the boundary elements that
have the apex as one of its vertices. This is not ideal, especially not
for coarse meshes, and may even be invalid. But, it is a step in the
right direction.