Add separate fiber computation function for single ventricle
All the region heuristics etc. done in `define_fibers` is not needed
when the mesh contains a single ventricle. Simply set the apicobasal
vector to be `grad_psi_ab` as usual, and set the transmural vector to
`phi_epi`. The angles are properly set with `alpha_w` and `beta_w`.

We don't really need to solve for `phi_lv` or `phi_rv` either, but...
yeah. For another day.